# docs/swagger.yaml
definitions:
...
  github.com_VinCPR_backend_api.blockInfoRequest:
    properties:
      block_name:
        type: string
      group_calendar:
        items:
          items:
            $ref: '#/definitions/github.com_VinCPR_backend_api.specialtyInfoRequest'
          type: array
        type: array
    type: object
...
  github.com_VinCPR_backend_api.designRotationRequest:
    properties:
      academic_year_name:
        type: string
      blocks:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.blockInfoRequest'
        type: array
      groups_per_block:
        type: integer
      number_of_period:
        type: integer
      periods:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.periodInfoRequest'
        type: array
      weeks_per_period:
        type: integer
    required:
    - academic_year_name
    - blocks
    - groups_per_block
    - number_of_period
    - periods
    - weeks_per_period
    type: object
...
  github.com_VinCPR_backend_api.hospitalInfoRequest:
    properties:
      hospital_name:
        type: string
      services:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.serviceInfoRequest'
        type: array
    type: object
...
  github.com_VinCPR_backend_api.periodInfoRequest:
    properties:
      period_name:
        type: string
      start_date:
        type: string
    required:
    - period_name
    - start_date
    type: object
...
  github.com_VinCPR_backend_api.serviceInfoRequest:
    properties:
      duration_in_week:
        type: integer
      service_name:
        type: string
    type: object
...
  github.com_VinCPR_backend_api.specialtyInfoRequest:
    properties:
      hospitals:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.hospitalInfoRequest'
        type: array
      specialty_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /rotation/design:
    post:
      consumes:
      - application/json
      description: provide templates of each period to auto generate calendar for
        all students
      parameters:
      - description: 'input required: academic year name, templates for each period'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.designRotationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
      summary: design clinical rotation
      tags:
      - ClinicalRotation
...
swagger: "2.0"